ai-dept-hub/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
chromadb==0.5.5
sentence-transformers==3.0.0
pdfminer.six==20231228
tiktoken==0.7.0
python-dotenv==1.0.1
bcrypt==4.1.3
